Интерфейс, enum и класс вынесены в отдельный файл
