+++ src/main/java/faang/school/godbless/transformation/Coordinates.java
@@ -1,0 +1,11 @@
+package faang.school.godbless.transformation;
+
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+
+@Getter
+@AllArgsConstructor
+class Coordinates {
+    private int x;
+    private int y;
+}
