+++ src/main/java/faang/school/godbless/transformation/FlipDirection.java
@@ -1,0 +1,6 @@
+package faang.school.godbless.transformation;
+
+public enum FlipDirection {
+    HORIZONTAL,
+    VERTICAL
+}
