+++ src/main/java/faang/school/godbless/transformation/MatrixTransformer.java
@@ -1,0 +1,6 @@
+package faang.school.godbless.transformation;
+
+@FunctionalInterface
+interface MatrixTransformer {
+    Coordinates transform(int x, int y);
+}
