+++ src/test/java/faang/school/godbless/transformation/TransformationTest.java
@@ -10,14 +10,14 @@ class TransformationTest {
     @Test
     void flipMatrixHorizontalTest() {
         int[][] exceptedMatrix = new int[][]{{3, 2, 1}, {6, 5, 4}, {9, 8, 7}};
-        int[][] flippedMatrix = Transformation.flipMatrix(matrix, Transformation.FlipDirection.HORIZONTAL);
+        int[][] flippedMatrix = Transformation.flipMatrix(matrix, FlipDirection.HORIZONTAL);
         assertArrayEquals(exceptedMatrix, flippedMatrix);
     }
 
     @Test
     void flipMatrixVerticalTest() {
         int[][] exceptedMatrix = new int[][]{{7, 8, 9}, {4, 5, 6}, {1, 2, 3}};
-        int[][] flippedMatrix = Transformation.flipMatrix(matrix, Transformation.FlipDirection.VERTICAL);
+        int[][] flippedMatrix = Transformation.flipMatrix(matrix, FlipDirection.VERTICAL);
         assertArrayEquals(exceptedMatrix, flippedMatrix);
     }
 }
